requirements.txt
PyYAML==6.0.1
selenium==4.15.2
requests==2.31.0
pandas==2.1.3
